Renamed tool to 'NDSRhythmMaker'
Added in fixes
Created an .exe
